Move utils method to quiz model
