+++ classes/Utils.php
@@ -29,6 +29,7 @@ class Utils {
 		$classes = array(
 			'Tutor\Models\CourseModel',
 			'Tutor\Models\LessonModel',
+			'Tutor\Models\QuizModel',
 			'Tutor\Models\WithdrawModel'
 		);
 
@@ -4487,34 +4488,6 @@ class Utils {
 	}
 
 	/**
-	 * @param int $question_id
-	 *
-	 * @return array|bool|object|void|null
-	 *
-	 * Get Quiz question by question id
-	 */
-	public function get_quiz_question_by_id( $question_id = 0 ) {
-		global $wpdb;
-
-		if ( $question_id ) {
-			$question = $wpdb->get_row(
-				$wpdb->prepare(
-					"SELECT *
-				FROM 	{$wpdb->prefix}tutor_quiz_questions
-				WHERE 	question_id = %d
-				LIMIT 0, 1;
-				",
-					$question_id
-				)
-			);
-
-			return $question;
-		}
-
-		return false;
-	}
-
-	/**
 	 * @param null $type
 	 *
 	 * @return array|mixed
@@ -4606,60 +4579,6 @@ class Utils {
 			return $answer_options;
 		}
 		return false;
-	}
-
-	/**
-	 * @param $quiz_id
-	 *
-	 * @return int
-	 *
-	 * Get the next question order ID
-	 *
-	 * @since v.1.0.0
-	 */
-	public function quiz_next_question_order_id( $quiz_id ) {
-		global $wpdb;
-
-		$last_order = (int) $wpdb->get_var(
-			$wpdb->prepare(
-				"SELECT MAX(question_order)
-			FROM 	{$wpdb->prefix}tutor_quiz_questions
-			WHERE 	quiz_id = %d ;
-			",
-				$quiz_id
-			)
-		);
-
-		return $last_order + 1;
-	}
-
-	/**
-	 * @param $quiz_id
-	 *
-	 * @return int
-	 *
-	 * new design quiz question
-	 * @since v.1.0.0
-	 */
-	public function quiz_next_question_id() {
-		global $wpdb;
-
-		$last_order = (int) $wpdb->get_var( "SELECT MAX(question_id) FROM {$wpdb->prefix}tutor_quiz_questions;" );
-		return $last_order + 1;
-	}
-
-	public function get_quiz_id_by_question( $question_id ) {
-		global $wpdb;
-		$quiz_id = $wpdb->get_var(
-			$wpdb->prepare(
-				"SELECT quiz_id
-			FROM 	{$wpdb->tutor_quiz_questions}
-			WHERE 	question_id = %d;
-			",
-				$question_id
-			)
-		);
-		return $quiz_id;
 	}
 
 	/**
@@ -5014,81 +4933,6 @@ class Utils {
 
 		return $answers;
 	}
-
-	/**
-	 * @param int $quiz_id
-	 * @param int $user_id
-	 *
-	 * @return array|bool|null|object
-	 *
-	 * Get all of the attempts by an user of a quiz
-	 *
-	 * @since v.1.0.0
-	 */
-
-	public function quiz_attempts( $quiz_id = 0, $user_id = 0 ) {
-		global $wpdb;
-
-		$quiz_id = $this->get_post_id( $quiz_id );
-		$user_id = $this->get_user_id( $user_id );
-
-		$attempts = $wpdb->get_results(
-			$wpdb->prepare(
-				"SELECT *
-			FROM 	{$wpdb->prefix}tutor_quiz_attempts
-			WHERE 	quiz_id = %d
-					AND user_id = %d
-					ORDER BY attempt_id  DESC
-			",
-				$quiz_id,
-				$user_id
-			)
-		);
-
-		if ( is_array( $attempts ) && count( $attempts ) ) {
-			return $attempts;
-		}
-
-		return false;
-	}
-
-	/**
-	 * @param int $quiz_id
-	 * @param int $user_id
-	 *
-	 * @return array|bool|null|object
-	 *
-	 * Get all ended attempts by an user of a quiz
-	 *
-	 * @since v.1.4.1
-	 */
-	public function quiz_ended_attempts( $quiz_id = 0, $user_id = 0 ) {
-		global $wpdb;
-
-		$quiz_id = $this->get_post_id( $quiz_id );
-		$user_id = $this->get_user_id( $user_id );
-
-		$attempts = $wpdb->get_results(
-			$wpdb->prepare(
-				"SELECT *
-			FROM 	{$wpdb->prefix}tutor_quiz_attempts
-			WHERE 	quiz_id = %d
-					AND user_id = %d
-					AND attempt_status != %s
-			",
-				$quiz_id,
-				$user_id,
-				'attempt_started'
-			)
-		);
-
-		if ( is_array( $attempts ) && count( $attempts ) ) {
-			return $attempts;
-		}
-
-		return false;
-	}
-
 
 	/**
 	 * @param int $user_id
